+++ Sudoku.Human/Core/Solver.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
+using Sudoku.Human.Core;
 
 namespace Kermalis.SudokuSolver.Core;
