+++ Sudoku.Human/Core/Solver_Methods.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using Sudoku.Human.Core;
 
 namespace Kermalis.SudokuSolver.Core;
